change vote data type

+++ src/user.go
@@ -7,7 +7,7 @@ import (
 type User struct {
 	Id   string
 	Name string
-	Vote int
+	Vote float64
 }
